requirements.txt
fastapi==0.116.1
passlib==1.7.4
psycopg2==2.9.10
pydantic==2.11.7
python_jose==3.5.0
SQLAlchemy==2.0.41
